ToDoList/serializers.py
from django.contrib.auth.models import User
from .models import Todo
from rest_framework import serializers

class TodoSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Todo
        fields = ['id','title', 'task', 'created_at']